% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lake_profile}
\alias{lake_profile}
\title{Lake Profile}
\format{
A data frame with 28 variables: \code{lake_name}, \code{depth}, \code{temp_c}, \code{do_mgl}, \code{do_sat_per}, \code{cond_umhos},
\code{light_sur_mmol}, \code{light_dep_mmol}, \code{ph}, \code{alk_mgl}, \code{turb_ntu},
\code{srp_mgl}, \code{tp_mgl}, \code{nh3_mgl}, \code{no3_mgl}, \code{tn_mgl},
\code{chla_ugl}, \code{olichospermum_nul}, \code{aphanizomenon_nul}, \code{microcystis_nul}, \code{ceratium_nul},
\code{nauplii_nul}, \code{bosmina_nul}, \code{calanoid_nul}, \code{cyclopoid_nul}, \code{chaoborus_nul},
\code{orgn_mgl}, and \code{light_level_per}.
}
\usage{
lake_profile
}
\description{
A dataset containing profile data from a Midwestern reservoir
}
\keyword{datasets}
