% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_percent.R
\name{light_percentage}
\alias{light_percentage}
\title{Calculate light percentage from surface at a discrete depth}
\usage{
light_percentage(light_at_depth, light_at_surface)
}
\arguments{
\item{light_at_depth}{Light measurement at discrete depth, mmol}

\item{light_at_surface}{Light measurement at surface, mmol}
}
\description{
Calculate light percentage from surface at a discrete depth
}
