% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_n.R
\name{organic_n}
\alias{organic_n}
\title{Calculate organic nitrogen concentration}
\usage{
organic_n(tn, no3, nh3)
}
\arguments{
\item{tn}{Total nitrogen concentration}

\item{no3}{Nitrate concentration}

\item{nh3}{Ammonia concentration}
}
\description{
Calculate organic nitrogen concentration
}
